requirements.txt
asgiref==3.5.0
Django==4.0.1
django-crispy-forms==1.14.0
django-environ==0.8.1
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
yapf==0.32.0
